Add option to build without Google deps
See https://github.com/iiordanov/remote-desktop-clients/issues/486
