+++ bVNC/src/withoutGoogle/java/com/iiordanov/bVNC/GoogleUtils.java
@@ -1,0 +1,9 @@
+package com.iiordanov.bVNC;
+
+import android.app.Activity;
+
+public class GoogleUtils {
+    public static void showRateAppDialog(Activity activity) {
+
+    }
+}
